Create a polymorphic association for profiles
Profiles are related to speaker profiles and event manager profiles
as information.

The content of the profiles controller is moved to the speaker profiles
controller.

There are currently a lot of broken things right now. These will be fixed
as the code is edited to accommodate this architectural shift.
