+++ app/models/event_manager_profile.rb
@@ -1,0 +1,3 @@
+class EventManagerProfile < ActiveRecord::Base
+  has_one :profile, as: :information, dependent: :destroy
+end
