+++ db/migrate/20150318203934_create_speaker_profiles.rb
@@ -1,0 +1,8 @@
+class CreateSpeakerProfiles < ActiveRecord::Migration
+  def change
+    create_table :speaker_profiles do |t|
+
+      t.timestamps null: false
+    end
+  end
+end
