+++ db/migrate/20150318203946_create_event_manager_profiles.rb
@@ -1,0 +1,8 @@
+class CreateEventManagerProfiles < ActiveRecord::Migration
+  def change
+    create_table :event_manager_profiles do |t|
+
+      t.timestamps null: false
+    end
+  end
+end
