+++ spec/factories/event_manager_profiles.rb
@@ -1,0 +1,6 @@
+FactoryGirl.define do
+  factory :event_manager_profile do
+    
+  end
+
+end
