+++ spec/factories/speaker_profiles.rb
@@ -1,0 +1,6 @@
+FactoryGirl.define do
+  factory :speaker_profile do
+    
+  end
+
+end
