+++ spec/models/event_manager_profile_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe EventManagerProfile, :type => :model do
+  pending "add some examples to (or delete) #{__FILE__}"
+end
